re-organizing server project structure to comply with cli
