+++ server/.golangci.yaml
