+++ server/internal/models/job/status.go
