+++ server/pkg/worker/stop.go
